section 11: implement authentication v2

+++ src/services/keyToken.service.js
@@ -32,7 +32,7 @@ class KeyTokenService {
 	}
 	
 	static findByUserId = async (userId) => {
-		return await keytokenModel.findOne({ user: userId }).lean()
+		return await keytokenModel.findOne({ user: userId })
 	}
 
 	static removeKeyById = async (id) => {
@@ -52,9 +52,7 @@ class KeyTokenService {
 	}
 
 	static deleteKeyById = async (userId) => {
-		return await keytokenModel.deleteOne({
-			user: userId
-		}).lean()
+		return await keytokenModel.deleteOne({ user: userId })
 	}
 }
